Valid nanosecond values are up to 1 second.
Initially, I had coded this to support leap seconds being represented as excess nanoseconds in the timestamp.
However, some days later I had forgotten whether some implementation really does it that way or whether I thought it would be a nice idea to to it like this.
Therefore I'm removing the possibility again; if forscript fails to record the nanoseconds when it is started during a leap second that wouldn't be the end of the world anyway.